Separated code from unit tests. Added unit tests related to player moves. 	new file:   game.py 	modified:   test_1.py
