+++ test_1.py
@@ -4,187 +4,9 @@ import pdb
 import sys
 import unittest
 from unittest import TestCase
+from game import TicTacToe, InvalidMove
 from bitarray import bitarray
-from exceptions import Exception
 
-row_wins = (bitarray('111000000'), bitarray('000111000'), bitarray('000000111'))
-col_wins = (bitarray('100100100'), bitarray('010010010'), bitarray('001001001'))
-diag_wins =(bitarray('100010001'), bitarray('001010100'))
-
-wins = row_wins + col_wins + diag_wins
-row_dict = {'a':0, 'b':3, 'c': 6}
-col_dict = {'1': 0, '2':2, '3':2}
-
-class InvalidMove(Exception):
-    def __init__(self, value):
-        self.value = value
-    def __str__(self):
-        return repr(self.value)
-
-class TicTacToe(object):
-    ''' This is the Game
-    '''
-    player_one = bitarray('000000000') # machine
-    player_two = bitarray('000000000') # opponent
-    blank_board =  bitarray('000000000')
-
-    def board(self):
-        return self.player_one | self.player_two
-
-    def reset(self):
-        self.player_one = bitarray('000000000') # machine
-        self.player_two = bitarray('000000000') # opponent
-
-    def parse_move(self, move):
-        try:
-            return row_dict[move[0]], col_dict[move[1]]
-        except:
-            raise InvalidMove(move)
-
-    def move(self, player, move):
-        row, col = self.parse_move(move)
-        if self.board()[row+col]==True :
-            return False
-        player[row+col]=True
-        return True
-
-    def legal_move(self, move):
-        row, col = self.parse_move(move)
-        try:
-            if self.board()[row+col]==False:
-                return True
-        except:
-            return False
-        return False
-
-    def is_there_a_winner(self):
-        for win in wins:
-            if (win & self.player_one)==win:
-                return self.player_one, win
-            if (win & self.player_two)==win:
-                return self.player_two, win
-        return False
-
-    def is_it_a_winner(self, player):
-        """Does this player have a winner"""
-        for win in wins:
-            if (win & player)==win:
-                return True
-        return False
-
-    def choose_move(self, player, offset=0):
-        '''simplest solution first, find the first empty cell'''
-        for a_index in range(0, len(player)):
-            position = (a_index + offset) % 9
-            if self.board()[position] is False:
-                player[position] = True
-                return True
-        return False #no moves left
-
-    def may_move_lead_to_win(self, move, player):
-        '''check if all possible rows, columns or  diags are blocked
-        '''
-        pass
-
-    def possible_moves(self, board=None):
-        """What moves are available."""
-        if board is None:
-            board = self.board()
-        choices = []
-        for position in range(0, 9):
-            if board[position]==False:
-                choices.append(position)
-        return choices
-
-    def is_there_a_winning_move(self, player, optional_block = -1):
-        """If there is such a move return it. Otherwise return false."""
-        t_play = player
-        pos = self.possible_moves()
-        for position in pos:
-            if position == optional_block:
-                continue
-            t_play[position] = True
-            if self.is_it_a_winner(t_play):
-                t_play[position] = False
-                return position
-            t_play[position] = False
-        return False
-
-    def move_that_results_with_two_winning_options(self, player):
-        """If there is such a move return it. Otherwise return false."""
-        choices = self.possible_moves()
-        count = 0
-        for choice in choices:
-            player[choice] = True
-            result = self.is_there_a_winning_move(player)
-            if result is not False:
-                count += 1
-                """Check for another winning play using this move"""
-                result = self.is_there_a_winning_move(player, result)
-                if result is not False:
-                    player[choice] = False
-                    return choice
-            player[choice] = False
-        return False
-
-    def wins_for_position(self, position):
-        p_wins = []
-        for w in wins:
-            if w[position] == True:
-                p_wins.append(w)
-        return p_wins
-
-    def most_winning_options(self, player):
-        """
-        player is opposing player
-        """
-        choices = self.possible_moves()
-        win_count_for_choice = {}
-        best = (choices[0], 0)
-        for choice in choices:
-            """find wins"""
-            wins_for_choice = self.wins_for_position(choice)
-            win_count = 0
-            for w in wins_for_choice:
-                if w & player==bitarray('000000000'):
-                    win_count += 1
-            if win_count>best[1]:
-                best = (choice, win_count)
-        return best
-
-    def make_the_best_move(self):
-        """Find the best possible move. Start by checking if there is a winning move for us"""
-        #pdb.set_trace()
-        result = self.is_there_a_winning_move(self.player_one)
-        if result is not False:
-            self.player_one[result] = True
-            return result
-        '''Check if opponent has a winning move, if so block it'''
-        result = self.is_there_a_winning_move(self.player_two)
-        if result is not False:
-            self.player_one[result] = True
-            return result
-        '''Is there a move that gives two immediate chances to win'''
-        result = self.move_that_results_with_two_winning_options(self.player_one)
-        if result is not False:
-            self.player_one[result] = True
-            return result
-        '''Does the opponent have a move that will give them two chance'''
-        result = self.move_that_results_with_two_winning_options(self.player_two)
-        if result is not False:
-            self.player_one[result] = True
-            return result
-        '''Find the move that gives us the most chances to win and our opponent the least '''
-        my_best_move, my_win_count = self.most_winning_options(self.player_two)
-        opponents_best_move, opp_win_count = self.most_winning_options(self.player_two)
-        if (my_win_count + opp_win_count)==0:
-            return False
-        if my_win_count>=opp_win_count:
-            self.player_one[my_best_move] = True
-            return (my_best_move, my_win_count)
-        else:
-            self.player_one[opponents_best_move] = True
-            return (opponents_best_move, opp_win_count)
 
 class TicTacToeTest(TestCase):
     '''Test the game '''
@@ -200,10 +22,57 @@ class TicTacToeTest(TestCase):
         del self.game
         self.game = None
 
+    def test_board(self):
+        self.game.player_one = bitarray('100100100')
+        self.game.player_two = bitarray('010010010')
+        self.assertEqual(bitarray('100100100')|bitarray('010010010'), self.game.board())
+
+    def test_reset(self):
+        self.game.player_one = bitarray('100100100')
+        self.game.player_two = bitarray('010010010')
+        self.assertEqual(bitarray('100100100')|bitarray('010010010'), self.game.board())
+        self.game.reset()
+        self.assertEqual(bitarray('000000000'), self.game.player_one)
+        self.assertEqual(bitarray('000000000'), self.game.player_two)
+
+    def test_parse(self):
+        try:
+            self.game.parse_move('ag4')
+        except InvalidMove:
+            pass
+        except:
+            print "Unexpected error:", sys.exc_info()[0]
+            raise
+        try:
+            self.game.parse_move('d3')
+            raise Exception('Should Error')
+        except InvalidMove:
+            pass
+        except:
+            print "Unexpected error:", sys.exc_info()[0]
+            raise
+        try:
+            self.game.parse_move('a0')
+            raise Exception('Should Error')
+        except InvalidMove:
+            pass
+        except:
+            raise Exception('Should Error')
+        self.assertEqual(self.game.parse_move('a1'),(0,0))
+
+    def test_move(self):
+        self.assertTrue(self.game.move(self.game.player_two, 'b1'))
+        self.assertFalse(self.game.move(self.game.player_one, 'b1'))
+        self.assertTrue(self.game.move(self.game.player_one, 'c1'))
+        self.assertEqual(self.game.board(), bitarray('000100100'))
+
     def test_possible_moves(self):
         #self.game.reset()
         #print self.game.player_one
         self.assertEqual(self.game.possible_moves(), [0,1,2,3,4,5,6,7,8])
+        self.assertTrue(self.game.move(self.game.player_two, 'b1'))
+        self.assertTrue(self.game.move(self.game.player_one, 'c1'))
+        self.assertEqual(self.game.possible_moves(), [0,1,2,4,5,7,8])
 
     def test_is_there_a_winning_move(self):
         self.game.player_two[0] = True
@@ -306,30 +175,12 @@ class TicTacToeTest(TestCase):
         #print self.game.is_it_a_winner(self.game.player_one)
         self.assertFalse(self.game.is_it_a_winner(self.game.player_one))
 
-    def test_parse(self):
-        try:
-            self.game.parse_move('d3')
-            raise Exception('Should Error')
-        except InvalidMove:
-            pass
-        except:
-            print "Unexpected error:", sys.exc_info()[0]
-            raise
-        try:
-            self.game.parse_move('a0')
-            raise Exception('Should Error')
-        except InvalidMove:
-            pass
-        except:
-            raise Exception('Should Error')
-        self.assertEqual(self.game.parse_move('a1'),(0,0))
-
     def test_recurse_attack(self):
         """
         Lets see if we can beat it
         """
         self.game.reset()
-        is_there_a_winner, game = self.recurse_attack(self.game)
+        is_there_a_winner, game = self._recurse_attack(self.game)
         if is_there_a_winner is not False:
             try:
                 print "The winner is"
@@ -340,10 +191,11 @@ class TicTacToeTest(TestCase):
                 print 'player_one', game.player_one
                 print 'player_two', game.player_two
 
-    def recurse_attack(self, tg):
+    def _recurse_attack(self, tg):
 #        print "recurse_level", self.recurse_level
         game = copy.deepcopy(tg)
         self.recurse_level += 1
+        self.assertLess(self.recurse_level, 10)
         #print "level", self.recurse_level
         #print "player_one here", self.game.player_one
         #print "player_two here", self.game.player_two
@@ -365,7 +217,7 @@ class TicTacToeTest(TestCase):
 #            print "no_win"
 #            print "player_one", game.player_one
 #            print "player_two", game.player_two
-            result, game = self.recurse_attack(game)
+            result, game = self._recurse_attack(game)
             self.assertFalse(result)
         return False, game
